% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{add.alpha}
\alias{add.alpha}
\title{add.alpha}
\usage{
add.alpha(col, alpha)
}
\arguments{
\item{col:}{A string vector of colour names or hex colour strings.}

\item{alpha:}{A number between 0 (solid) and 1 (fully transparent).}
}
\value{
A string vector of modified hex colour strings.
}
\description{
\code{add.alpha} modifies a vector of named ('firebrick') or hexdecimal 
('#B22222') color specifications to add an alpha (transparency) channel.  
Inspired by https://gist.github.com/mages/5339689
}
\examples{
require(RColorBrewer)
pal <- add.alpha(brewer.pal(5, 'Set1'), 0.5)
plot(rnorm(100), rnorm(100), pch=21, col='black', bg=pal, cex=10)

}
