% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{add.scalebar}
\alias{add.scalebar}
\title{add.scalebar
Add a scale bar to the plot.}
\usage{
add.scalebar(obj, len = 1, x0 = NA, y0 = NA, lwd = 2, ...)
}
\arguments{
\item{obj:}{`phyloLayout` S3 object generated by `tree.layout`}

\item{len:}{length of scale bar to draw (defaults to 1)}

\item{x0:}{(optional) override default horizontal placement of bar}

\item{y0:}{(optional) override default vertical placement of bar.
Note both x and y need to be specified to override the default.}

\item{...:}{additional graphical parameters to pass to `segments`, 
such as `cex`, `col` and `lwd`.}
}
\description{
add.scalebar
Add a scale bar to the plot.
}
\examples{
require(ape)
phy <- rcoal(30)
Y <- tree.layout(phy, type='r')
plot(Y)
add.scalebar(Y, len=1)
}
