% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{as.phyloData}
\alias{as.phyloData}
\title{as.phyloData}
\usage{
as.phyloData(phy, unscaled = FALSE)
}
\arguments{
\item{phy:}{an S3 object of class `phylo` (`ape` package), which 
must be a rooted tree.}
}
\value{
an S3 object of class `phyloData`
}
\description{
Convert a `phylo` object into a data frame of edge information and 
a vector of node information.
}
\examples{
require(ape)
phy <- rtree(20)
as.phyloData(phy)

}
