% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{blend.colors}
\alias{blend.colors}
\title{blend.colors}
\usage{
blend.colors(col)
}
\arguments{
\item{col:}{a vector of named or hexadecimal color specifications}
}
\value{
a hex color string, including an alpha channel
}
\description{
Average a vector of colours in RGBA colour space.
}
\examples{
blend.colors(c('red', 'blue', 'green'))

}
