% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookplot.R
\name{bookplot}
\alias{bookplot}
\title{bookplot}
\usage{
bookplot(
  x,
  y = NA,
  space = 1,
  edges = T,
  border = NA,
  col = NA,
  lwd = 1,
  xaxt = "s",
  axis.srt = 90,
  axis.offset = 0.05,
  cex.axis = 1,
  edge.col = "grey",
  edge.lwd = 0.5,
  legend = NA
)
}
\arguments{
\item{x:}{factor, independent variable}

\item{y:}{factor, dependent variable}

\item{space:}{amount of space between stacked bars}

\item{edges:}{if TRUE, draw line segments connecting adjacent rectangles 
of the same variable.}

\item{border:}{colour for rectangle borders; defaults to NA (no border)}

\item{col:}{colour(s) to fill rectangles}
}
\description{
Similar to graphics::spineplot, a bookplot visualizes the contingency table 
of two categorical variables (factors).  The width of each stack is 
scaled to the square root of the marginal total, and the height of each 
is scaled such that the area of each rectangle in the stack is proportional 
to the corresponding count.
}
\examples{


}
