\name{co2.emissions}
\docType{data}
\alias{co2.emissions}
\title{CO2 emissions per capita by nation}
\description{
  This is a small data frame generated from a CSV file published at https://cdiac.ess-dive.lbl.gov/trends/emis/tre_coun.html. 
  It is similar to the `CO2_emissions` data set used by @clauswilke to demonstrate the use of slopegraphs in Wilke (2019).
}
\references{
Wilke CO. Fundamentals of Data Visualization: A Primer on Making Informative and Compelling Figures. O'Reilly Media; 2019 Mar 18.
}
\keyword{datasets}
