% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{count.tips}
\alias{count.tips}
\title{count.tips}
\usage{
count.tips(phy)
}
\arguments{
\item{phy:}{an S3 object of class `phylo` (`ape`` package)}
}
\value{
numeric vector of length Nnode(phy), sorted in preorder
}
\description{
Calculate the number of tips for every internal node in the tree.
}
