% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{.climb.tree}
\alias{.climb.tree}
\title{Traverse tree from given node (tip) to root, return 
vector of indices.}
\usage{
.climb.tree(node, edges, result = c())
}
\description{
Traverse tree from given node (tip) to root, return 
vector of indices.
}
\keyword{internal}
