% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{.layout.equalangle}
\alias{.layout.equalangle}
\title{layout.equalangle}
\usage{
.layout.equalangle(phy, unscaled)
}
\description{
Genearte the coordinates for an unrooted layout of a phylogenetic
 tree using Joe Felsenstein's equal-angle algorithm.
 
 @param phy:  an S3 object of class `phylo`
 
 @return S3 object of class `phyloLayout`
 @keywords internal
}
