% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{.layout.radial}
\alias{.layout.radial}
\title{layout.radial}
\usage{
.layout.radial(phy, unscaled)
}
\arguments{
\item{phy:}{an S3 object of class `phylo`}
}
\value{
S3 object of class `phyloLayout`
}
\description{
Generate the coordinates for a radial layout of a phylogenetic tree
within a circular region.
}
\keyword{internal}
