% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{.layout.rect}
\alias{.layout.rect}
\title{.layout.rect}
\usage{
.layout.rect(phy, unscaled = FALSE, slanted = FALSE)
}
\description{
Algorithm to generate a "rectangular" (left to right) layout of a 
phylogenetic tree, with the root at the left and tips on the right.
}
\keyword{internal}
