% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{.recursive.equalAngles}
\alias{.recursive.equalAngles}
\title{.recursive.equalAngles}
\usage{
.recursive.equalAngles(index, pd)
}
\arguments{
\item{index:}{row number of current node in data frame}

\item{pd:}{S3 object of class `phyloData``}
}
\value{
`phyloData` object with updated `nodes` data frame
}
\description{
A recursive function for calculating Felsenstein's equal-angles algorithm
for unrooted tree layouts.
}
\keyword{internal}
