% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{.rotate.label}
\alias{.rotate.label}
\title{.rotate.label
Adjust the string rotation of node labels for unrooted
or radial trees so they are not upside-down.}
\usage{
.rotate.label(node)
}
\arguments{
\item{node:}{named vector, a row from the nodes data frame of a 
`phyloLayout` object.}
}
\value{
a named vector with updated `x`, `y`, `angle` and `label` values
}
\description{
.rotate.label
Adjust the string rotation of node labels for unrooted
or radial trees so they are not upside-down.
}
\keyword{internal}
