% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{draw.branch}
\alias{draw.branch}
\title{draw.branch}
\usage{
draw.branch(layout, tips, col = "red", ...)
}
\arguments{
\item{layout:}{an S3 object of class `phyloLayout`}

\item{tips:}{a character vector of tip labels}

\item{col:}{color for drawing branch, defaults to red}

\item{...:}{additional arguments passed to `segments`}
}
\description{
Locate the common ancestor given a vector of tip labels and 
draw the corresponding branch for a given tree layout.
}
