% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{draw.clade}
\alias{draw.clade}
\title{draw.clade}
\usage{
draw.clade(layout, tips, col = "red", ...)
}
\arguments{
\item{layout:}{an S3 object of class `phyloLayout`}

\item{tips:}{a character vector of tip labels}
}
\description{
Locate common ancestor in the tree given a set of tip labels
and draw the subtree (clade) of all descendants.
}
