% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{draw.guidelines}
\alias{draw.guidelines}
\title{draw.guidelines}
\usage{
draw.guidelines(obj, lty = 3, ...)
}
\arguments{
\item{obj:}{S3 object of class `phyloLayout`}

\item{lty:}{line style, defaults to 3 (dotted)}

\item{...:}{other graphical parameters to pass to `segments`,
such as `col` or `lwd`}
}
\description{
Draws segments from the tips of the tree out to the furthest
tip label.  This should only be used in conjunction with 
text.phyloLayout(align=TRUE), or else the segments will be drawn
over the tip labels.
}
