% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{draw.stack}
\alias{draw.stack}
\title{draw.stack}
\usage{
draw.stack(x, xoffset, yoffset = 0, width = 1, space = 0.1, col = NA, ...)
}
\arguments{
\item{x:}{numeric vector or matrix of non-negative values}

\item{offset:}{horizontal location of bars relative to origin}

\item{width:}{}
}
\description{
Draw stacked bars as in barplot(), but with arbitrary location
on x-axis.
TODO: extend for circular layouts using draw.arc
}
