% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{gg.rainbow}
\alias{gg.rainbow}
\title{gg.rainbow}
\usage{
gg.rainbow(n, c = 100, l = 65, alpha = 1)
}
\arguments{
\item{n:}{number of colours to generate}

\item{c:}{chroma, radial component of hue colour space (saturation)}

\item{l:}{luminance in range [0, 100], vertical component of 
hue colour space (converges to black as l approaches 0)}
}
\value{
vector of colour specification strings
}
\description{
Mimic default colour palette of ggplot2.
Credit: John Colby, https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
