% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{image.phyloLayout}
\alias{image.phyloLayout}
\title{image}
\usage{
\method{image}{phyloLayout}(obj, z, col)
}
\arguments{
\item{obj:}{an S3 object of class `phyloLayout`}

\item{z:}{matrix, data to annotate}

\item{col:}{a vector of colors}
}
\description{
Generic function for drawing a grid of coloured or grey-scale 
rectangles corresponding to values in a matrix `z`.  Rows in 
the matrix are assumed to correspond to tips of the tree.
Only rectangular or slanted tree layouts are supported.  If 
you want to annotate a circular (radial) layout, use `ringplot`.
}
