% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{label.phyloLayout}
\alias{label.phyloLayout}
\title{label.phyloLayout}
\usage{
label.phyloLayout(obj, label = "t", cex.lab = 1, ...)
}
\arguments{
\item{obj:}{an S3 object of class `phyloLayout`}

\item{label:}{Specifies which set of nodes are labeled with text.
\itemize{
  \item{'t'}{Tip labels only (default).}
  \item{'i'}{Internal node labels only.}
  \item{'b'}{Both tip and internal node labels.}
}}

\item{cex.lab:}{character expansion factor for text (default 1)}

\item{...:}{additional graphical parameters passed to `text`}
}
\description{
Draw tip and/or internal node labels on the phylogenetic tree.
}
