% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{lines.phyloLayout}
\alias{lines.phyloLayout}
\title{lines.phyloLayout}
\usage{
\method{lines}{phyloLayout}(obj, ...)
}
\arguments{
\item{obj:}{an S3 object of class `phyloLayout`}

\item{...:}{additional graphical parameters to pass to `segments`
and `draw.arc`}
}
\description{
Generic S3 method adapted for drawing the line segments that 
comprise a phylogenetic tree.
}
