% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{points.phyloLayout}
\alias{points.phyloLayout}
\title{points.phyloData}
\usage{
\method{points}{phyloLayout}(obj, ...)
}
\arguments{
\item{obj:}{S3 object of class `phyloLayout`}

\item{...:}{additional graphical parameters to pass to 
points()}
}
\description{
A generic function to add points to the plot.  This is 
customized for phyloLayout S3 objects that provide the x-
and y- coordinates for tips and internal nodes.
}
\examples{
# the structSI `phylo` object was generated by simulating
# an epidemic within a structured population
l <- tree.layout(structSI, type='r') 
plot(l)

# highlight nodes that represent transmission or migration events
cex <- as.integer(structSI$event.type \%in\% c('transmission', 'migration'))
bg <- ifelse(structSI$event.type=='transmission', 'dodgerblue', 'salmon')
points(l, cex=cex, pch=21, bg=bg)

}
