% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringplot.R
\name{polarplot}
\alias{polarplot}
\title{polarplot}
\usage{
polarplot(obj, r = 0, theta = 0.5 * pi, decay = 0.8, col = NA, ...)
}
\arguments{
\item{obj:}{a numeric vector, matrix or table of frequency data}

\item{r:}{radius for inner circle, defaults to 0}

\item{theta:}{rotation offset for plot in radians, defaults to pi/2}

\item{decay:}{wedges become "squished" with increasing distance from the
origin - 'decay' is a factor between 0 and 1 that progressively
reduces the arc length of wedges.  Defaults to 1 (no decay).}

\item{col:}{a vector of colour strings}

\item{...:}{additional arguments passed to the \code{plot} function.}
}
\description{
A circular plot where the circle is partitioned by radii at equal angles
and the areas of the resulting sectors are proportional to the data.
}
\examples{
require(RColorBrewer)
require(HistData)
pal <- brewer.pal(3, 'Pastel2')
ng <- subset(Nightingale, Year==1855, c('Other.rate', 'Wounds.rate', 'Disease.rate'))
row.names(ng) <- Nightingale$Month[Nightingale$Year==1855]
par(mar=rep(0,4))
polarplot(as.matrix(ng), decay=1, theta=1.1*pi, col=pal)
title('Causes of mortality in British army, Crimean War (1855)', 
      font.main=1, family='Garamond', line=-2, cex.main=1.1)

}
