% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print.phyloData}
\alias{print.phyloData}
\title{print.phyloData}
\usage{
\method{print}{phyloData}(obj)
}
\arguments{
\item{obj:}{S3 object of class `phyloData`}
}
\description{
Generic function to display object.
}
