% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print.phyloLayout}
\alias{print.phyloLayout}
\title{print.phyloLayout}
\usage{
\method{print}{phyloLayout}(obj)
}
\arguments{
\item{obj:}{S3 object of class `phyloLayout`}
}
\description{
Generic function to display object.
}
