% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeplot.R
\name{ridgeplot}
\alias{ridgeplot}
\title{ridgeplot}
\usage{
ridgeplot(
  x,
  xlim = NA,
  labels = NA,
  yaxt = "s",
  xlab = NA,
  ylab = NA,
  step = 0.2,
  col = NA,
  fill = NA,
  lwd = 1,
  density.args = list(),
  add.grid = F,
  grid.args = list(),
  extend.lines = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{x:}{a list of numeric vectors, such as produced by calling split().}

\item{xlim:}{by default, automatically set to range of horizontal axis.}

\item{labels:}{override names in list 'x'}

\item{xlab:}{string to label x-axis (default NA)}

\item{ylab:}{string to label y-axis (default NA)}

\item{step:}{numeric value to determine spread and ordering of densities.
Negative values cause reversed ordering.}

\item{col:}{character vector of R colours for lines and borders.  Will recycle 
colours if there are fewer than the number of groups in 'x'.}

\item{fill:}{character vector of R colours for filling densities.  Will recycle
colours if there are fewer than the number of groups in 'x'.}

\item{lwd:}{line width for drawing density curves and line segments.}

\item{density.args:}{list of optional arguments for density().}

\item{add.grid:}{if TRUE, call add.grid() before drawing densities}

\item{grid.args:}{list of optional arguments for add.grid()}

\item{extend.lines:}{whether to draw densities across the horizontal range of plot.
Defaults to TRUE.}

\item{add:}{if TRUE, add the series of densities to existing plot.}

\item{...:}{additional arguments for plot()}
}
\description{
\code{ridgeplot} generates stacked density plots (aka. "ridge plots") in 
the style of the package \code{ggridge}, but using only base R graphics.
}
\examples{
ridgeplot(split(iris$Sepal.Length, iris$Species))

# reverse ordering, use solid colour lines and modify kernel density
ridgeplot(split(iris$Sepal.Length, iris$Species), step=-0.2, 
          col=c('firebrick', 'steelblue', 'olivedrab'), lwd=5, 
                    density.args=list(n=20, kernel='rectangular', adjust=0.5))
                    
# fill densities with translucent colours
require(RColorBrewer)
pal <- add.alpha(brewer.pal(3, 'Set1'), 0.5)
ridgeplot(split(iris$Sepal.Length, iris$Species), step=0.4, col='white', 
fill=pal, lwd=2, xlab='Sepal length', cex.lab=1.2)

# passing arguments to add.grid()
ridgeplot(split(iris$Sepal.Length, iris$Species), step=0.4, col='white', 
add.grid=T, grid.args=list(mode='y', bg.col='black', fg.col='grey', 
lwd.major=0.5, lwd.minor=0))

}
