\name{stjohns}
\docType{data}
\alias{stjohns}
\title{Wind speed and direction at St. John's Interational Airport.}
\description{
This data frame contains wind direction (10s deg) and speed (km/h) at the St. John's International Airport in Newfoundland, Canada, throughout May 2016.
Wind blowing from the south has a direction of 180 degrees (18.0 10s deg).

These data were obtained from the Government of Canada Historical Climate Data website: http://climate.weather.gc.ca/climate_data/hourly_data_e.html?StationID=50089&Month=5&Day=1&Year=2016
}
\keyword{datasets}
