% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{text.phyloLayout}
\alias{text.phyloLayout}
\title{text.phyloLayout}
\usage{
\method{text}{phyloLayout}(obj, label = "t", align = FALSE, cex.lab = 1, offset = 0, ...)
}
\arguments{
\item{obj:}{an S3 object of class `phyloLayout`}

\item{label:}{Specifies which set of nodes are labeled with text.
\itemize{
  \item{'t'}{Tip labels only (default).}
  \item{'i'}{Internal node labels only.}
  \item{'b'}{Both tip and internal node labels.}
}}

\item{align:}{if TRUE, all tip labels are drawn at maximum value}

\item{cex.lab:}{character expansion factor for text}

\item{...:}{additional graphical parameters passed to `text`}
}
\description{
Draw tip and/or internal node labels on the phylogenetic tree.
}
