% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree.layout}
\alias{tree.layout}
\title{tree.layout}
\usage{
tree.layout(phy, type = "r", unscaled = FALSE)
}
\arguments{
\item{phy:}{an S3 object of class `phylo``}

\item{type:}{what type of layout algorithm to employ:
\itemize {
  \item{"r"}{Rectangular layout}
  \item{"s"}{Slanted (triangular) layout}
  \item{"u"}{Unrooted (equal-angle) layout}
  \item{"o"}{Circular (radial) layout}
}}

\item{unscaled:}{if TRUE, return a cladogram layout}
}
\value{
S3 object of class `phyloLayout`
}
\description{
Generate the coordinates for the nodes and edges of a 
phylogenetic tree (ape:phylo object) using one of several
layout algorithms.  Used before calling the generic plot
functions (plot, points).
}
\examples{
require(ape)
phy <- rcoal(20)
lay <- tree.layout(phy, 'r')
head(lay$edges)

}
