% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{unroot}
\alias{unroot}
\title{unroot}
\usage{
unroot(obj)
}
\description{
If the `phylo` object contains non-default attributes for nodes
or edges, then unrooting breaks the alignment between these 
attributes and the tree.
}
