% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{draw.arc}
\alias{draw.arc}
\title{draw.arc}
\usage{
draw.arc(x, y, r0, r1, theta0, theta1, border = NULL, col = NA,
  lty = par("lty"), n = 64)
}
\arguments{
\item{x:}{x-coordinate of origin}

\item{y:}{y-coordinate of origin}

\item{r0:}{radius to inner arc}

\item{r1:}{radius to outer arc}

\item{theta0:}{start angle of arc}

\item{theta1:}{end angle of arc}

\item{border:}{colour for line around polygon border}

\item{col:}{colour for polygon fill}

\item{lty:}{line type parameter}

\item{n:}{number of points to interpolate curves}
}
\description{
\code{draw.arc} draws an annular sector around origin (x,y)
as a polygon
}
\examples{
# create plot region
par(mar=c(5,5,1,1))
plot(NA, xlim=c(-1, 1), ylim=c(-1, 1))
draw.arc(0, 0, 0.5, 0.8, 0.7*pi, 1.1*pi, col='steelblue')

}
