% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{image.phyloLayout}
\alias{image.phyloLayout}
\title{image}
\usage{
\method{image}{phyloLayout}(obj, z, xlim = NA, col = NA, border = "white", xaxt = "y", ...)
}
\arguments{
\item{obj:}{an S3 object of class `phyloLayout`}

\item{z:}{matrix, data to annotate tips in order of *nodes*; this will be 
recast as a factor, and then an integer vector.  If your input 
`z` is a numeric (continuous-valued) vector, then you should run 
`cut(z)` to discretize the distribution.}

\item{xlim:}{limits (x1, x2) of grid relative to current plot device.
For a radial tree layout, these correspond to the inner and 
outer radii.  You may need to use trial-and-error to find 
a good set of boundaries.  Note this function will call 
`xpd=NA` to permit drawing in margins.
Defaults to `max(x)+0.01*range(x)` to `max(x)+0.06*range(x)`}

\item{col:}{a vector of colours that maps to factor levels in `z`}

\item{border:}{colour for border of rectangles in grid}

\item{xaxt:}{if 'n', suppress drawing of axis and labels}
}
\description{
Generic function for drawing a grid of coloured or grey-scale 
rectangles corresponding to values in a matrix `z`.  Rows in 
the matrix are assumed to correspond to tips of the tree.
}
