% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopegraph.R
\name{slopegraph}
\alias{slopegraph}
\title{slopegraph}
\usage{
slopegraph(x, y, names = NA, labels = NA, xlab = "Groups",
  ylab = NA, type = "b", colorize = F, pal = c("firebrick",
  "steelblue"), shim = 0.2, cex.labels = 0.8, add.grid = F,
  grid.args = list(), ...)
}
\arguments{
\item{x:}{numeric vector of values to plot on the left}

\item{y:}{numeric vector of values to plot on the right}

\item{groups:}{optional names for labeling the x-axis}

\item{labels:}{character vector of labels to annotate slopes}

\item{xlab:}{label for x-axis, as in generic \code{plot()}. Defaults to 'Groups'.}

\item{ylab:}{label for y-axis, as in generic \code{plot()}. Defaults to \code{NA}.}

\item{type:}{'b' displays both points and lines; 'l' displays lines only.}

\item{colorize:}{use line colours to emphasize slope}

\item{pal:}{vector of two colours for annotating slopes down and up, respectively.}

\item{shim:}{width of horizontal padding to left and right of plot}

\item{add.grid:}{if TRUE, calls the function \code{ggfree::add.grid}}

\item{grid.args:}{list of optional arguments for \code{add.grid()}}

\item{...:}{additional arguments to pass to \code{lines()}, *e.g.*, \code{lwd}}
}
\description{
\code{slopegraph} displays paired sample data as two sets of points connected 
by line segments as a visualization of overall trends.
}
\examples{
slopegraph(iris$Sepal.Width, iris$Petal.Length, colorize=T, add.grid=T, 
pch=21, bg='white')

}
