% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackplot.R
\name{stackplot}
\alias{stackplot}
\title{stackplot}
\usage{
stackplot(
  obj,
  x = NA,
  freq = TRUE,
  center = FALSE,
  spline = FALSE,
  density = NA,
  angle = NA,
  col = NA,
  border = NULL,
  lwd = 1,
  xlim = NA,
  ...
)
}
\arguments{
\item{obj:}{an object of class 'matrix' or 'table'.  Rows are assumed 
to correspond to repeated measures, and columns to 
different groups.}

\item{x:}{optional, customize the horizontal location of data
points (i.e., counts in `obj`)}

\item{freq:}{if FALSE, normalize to relative frequencies}

\item{center:}{if TRUE, normalize frequencies to the mean for 
centered baseline}

\item{spline:}{if TRUE, interpolate data points to draw areas with 
smoothed trendlines}

\item{density:}{numeric, density of shading lines for each area;
values recycled as necessary}

\item{angle:}{numeric, angle of shading lines for each area;
values recycled as necessary}

\item{col:}{character, colour specification strings for filling
areas, recycled as necessary}

\item{border:}{border color for polygon areas}

\item{lwd:}{line weight for polygon borders}

\item{xlim:}{optional, allows user to override default horizontal
plot range}

\item{...:}{additional arguments to pass to `plot` function}
}
\description{
Generates a stacked area plot of the absolute or relative
frequencies of groups over time (or some other discrete or 
continuous variable).
}
\details{
Credit for streamplot
https://gist.github.com/menugget/7864454
}
\examples{
stackplot(EuStockMarkets, xlab='Days (1991-1998)', 
  ylab='Daily Closing Price', col=gg.rainbow(4), border=NA, 
  bty='n')

# a centered stackplot with shading 
last.mar <- par()$mar
par(mar=rep(1,4))
stackplot(EuStockMarkets, center=TRUE, density=seq(0, 60, 20), 
  angle=30, border='black', bty='n', xaxt='n', yaxt='n', lwd=0.5)
par(mar=last.mar)

# diversity of SARS-CoV-2 in 2020
par(mar=c(1,1,3,1))
set.seed(1); pal <- gg.rainbow(ncol(cov2))[sample.int(ncol(cov2))]
stackplot(cov2, col=pal, center=T, bty='n', xaxt='n', yaxt='n',
xlab=NA, ylab=NA, spline=T, border='black', lwd=0.2, xlim=c(10, 20))
axis(side=3, at=seq(10, 20, 2), line=1, cex.axis=0.8, col='grey60')
par(mar=last.mar)

}
